+++ src/api/browserview/browserview.ts
@@ -1,8 +1,8 @@
 import { WebContents } from '../webcontents/webcontents';
-import { BaseEventMap } from '../events/base';
 import Transport from '../../transport/transport';
 import { Identity } from '../../identity';
 import { Base } from '../base';
+import { ViewEvents } from '../events/browserview';
 export interface AutoResizeOptions {
     /**
      * If true, the view's width will grow and shrink together with the window. false
@@ -51,7 +51,7 @@ export class BrowserViewModule extends Base {
     }
 }
 
-export class BrowserView extends WebContents<BaseEventMap> {
+export class BrowserView extends WebContents<ViewEvents> {
     constructor(wire: Transport, public identity: Identity) {
         super(wire, identity, 'browserview');
     }
